+++ src/app/(protected)/orders/useInvalidateQueries.ts
@@ -1,0 +1,5 @@
+"use client";
+
+export const useInvalidateQueries = () => {
+    return () => {};
+};
